added Relations + ecryption

+++ records/views.py
@@ -4,7 +4,8 @@ from rest_framework.response import Response
 from django.shortcuts import render,redirect,  get_object_or_404
 from django.contrib.auth import login, authenticate, logout
 from django.contrib.auth.forms import AuthenticationForm
-from django.http import HttpResponse
+from django.contrib.auth.models import User
+from django.http import HttpResponse, Http404, HttpResponseForbidden
 from .models import HealthRecord
 from .serializers import HealthRecordSerializer
 from .blockchain import create_record, grant_access, revoke_access, get_record
@@ -15,7 +16,10 @@ from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from .forms import RegisterForm
 import ipfshttpclient2
-import mimetypes
+import mimetypes, os
+from .signature import generate_ecdsa_keys, sign_file,  verify_signature
+from django.contrib import messages
+
 
 def register(request):
     if request.method == 'POST':
@@ -84,42 +88,74 @@ class HealthRecordViewSet(viewsets.ModelViewSet):
 
 @csrf_exempt
 def upload_file(request):
+    if not request.user.is_superuser:
+        return HttpResponseForbidden("Only admins can upload records.")
     health_record = HealthRecord()
+    users = User.objects.all()
     if request.method == 'POST':
         file = request.FILES['file']
         file_name = file.name
         file_content = file.read()
-        file_record_type = request.POST['record_type']
+        patient_id = request.POST['patient_id']
+        patient = User.objects.get(id=patient_id)
+        doctor_id = request.POST['doctor_id']
+        doctor = User.objects.get(id=doctor)
         
         # Save the file temporarily
         temp_file_path = f"{settings.MEDIA_ROOT}/temp_{file_name}"
         with open(temp_file_path, 'wb') as temp_file:
             temp_file.write(file_content)
+
+        pem_private, pem_public = generate_ecdsa_keys()
+
+        signature = sign_file(pem_private, temp_file_path)
             
         cid = ipfs_api.publish(temp_file_path)
         ipns_key = f"HealthRecord_{file_name}"
         ipfs_api.create_ipns_record(ipns_key)
         ipfs_api.update_ipns_record(ipns_key, temp_file_path)
-        import os
-        os.remove(temp_file_path)
+        
         health_record = HealthRecord.objects.create(
-                patient=request.user,
-                record_name=file_name,
+                patient=patient,
+                doctor=doctor,
+                record_name=file_name,  
                 record_id=cid,
                 ipfs_hash=ipns_key,
+                signature=signature.hex()
             )
-        return HttpResponse(f"File uploaded to IPFS. CID: {cid}, IPNS Key: {ipns_key}")
-    return render(request, 'upload.html')
+        
+        os.remove(temp_file_path) # remvoe file from temp folder
+
+        messages.success(request, f"File uploaded to IPFS. CID: {cid}, IPNS Key: {ipns_key}")
+        return redirect('upload_file')
+    return render(request, 'upload.html', {'users': users})
 
 def retrieve_file(request, cid):
     try:
         client = ipfshttpclient2.connect()
         file_data = client.cat(cid)
+        
+        # Fetching health record and signature
+        health_record = HealthRecord.objects.get(record_id=cid)
+        signature = bytes.fromhex(health_record.signature)
+
+        temp_file_path = f"{settings.MEDIA_ROOT}/temp_{cid}"
+        with open(temp_file_path, 'wb') as temp_file:
+            temp_file.write(file_data)
+        
+        is_verified = verify_signature(health_record.patient.profile.public_key, temp_file_path, signature)
+        os.remove(temp_file_path)
+
+        if not is_verified:
+            return HttpResponse("Signature verification failed", status=403)
+        
         file_extension = cid.split('.')[-1]
         content_type = mimetypes.guess_type(f"file.{file_extension}")[0] or 'application/octet-stream'
         response = HttpResponse(file_data, content_type=content_type)
         response['Content-Disposition'] = 'inline' 
         return response
+    except HealthRecord.DoesNotExist:
+        return Http404(f"HealthRecord with CID {cid} not found.")
     except Exception as e:
         return HttpResponse(f"Error retrieving file: {str(e)}", status=404)
     
@@ -141,5 +177,9 @@ def record_detail(request, pk):
 
 @login_required(login_url='/login/')
 def view_records(request):
-    records = HealthRecord.objects.filter(patient=request.user)
+    if request.user.is_superuser:  # Admin access
+        records = HealthRecord.objects.all()    
+    else:
+        records = HealthRecord.objects.filter(patient=request.user) | HealthRecord.objects.filter(doctor=request.user)
+
     return render(request, 'view_records.html', {'records': records})
